backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3

# Database
PyMySQL==1.1.0
cryptography==41.0.7

# Validation & Serialization
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Security
bcrypt==4.1.1
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# ML & Data Science (for chatbot forecasting)
numpy==1.24.3

# Development
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
